requirements.txt
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
docutils==0.16
flit==3.0.0
flit-core==3.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
numpy==1.19.5
pandas==1.1.5
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.5
requests==2.25.1
six==1.15.0
twstock==1.3.1
urllib3==1.26.2
Werkzeug==1.0.1
flask
gunicorn
line-bot-sdk
